+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Bump to eslint-config-sonarqube@0.6.1
+- SC-2266 Add a `setRequestOptions`  to the initialisation process
 
 ## 1.0.15