+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.0.17
+
 - SC-2266 onVersionChange request option must return a boolean to allow the request to proceed or not
 
 ## 1.0.16